requirements.txt
streamlit==1.26.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
joblib==1.3.2
transformers
safetensors
torch
